[FUNC][CFG][R] Added API for currency conversion
Added POST API with path "/convert" to the ExchangeController.

Added a new model CurrencyConversionData. The model contains source
currency field, converted currency and amount. The amount is validated
to be greater than 0.

Added 'spring-boot-starter-validation' dependency for validator
implementation.

Updated ApiError to contain a list of errors instead of only one