[R] Refactored conversion logic
The ExchangeService now has a convert method. The merhod converts the
given source currency amount to the target currency and saves the
conversion transaction to the database.

The ExchangeService::getRate method no longer has a saveTransaction
parameter and never saves the transaction to the database.

The ConversionEntity now has convertedAmount field. The amount field was
renamed to source amount.

Added a ConversionResult model which contains conversion transaction ID
and converted value.

The ExchangeController::convert API now returns ConversionResult.