[FUNC] Added APIs to retrieve transaction history
Added /get-transaction API to ExchangeController. The API accepts a
transaction ID and returns the transaction details.

Added /get-transactions API to ExchangeController. The API accepts a
date, page number and page size and returns requested page and size of
transactions for the given day.

Updaetd application-template.properties. Added the following date
formatting property "spring.mvc.format.date=yyyy-MM-dd".

Added a Conversion DTO for the ConversionEntity.

Added a toConversion method in the ConversionEntity which returns a
Conversion DTO initialized with the entity values.

Added a method in ConversionRepository which retrieves all conversions
for a given date-time interval in milliseconds.

Added getConversion and conversionHistoryByDate methods to
ExchangeService. The first method retrieves a conversion by a given ID.
The second method returns a page of all conversions for a given day.

Added InvalidPageNumberException and InvalidPageSizeException.